fix: refactor and restructure code
